+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+FEATURES:
+* Add beta endpoints `ApplyToProjects`  and `RemoveFromProjects` to `VariableSets`.  Applying a variable set to a project will apply that variable set to all current and future workspaces in that project. 
+* Add beta endpoint `ListForProject` to `VariableSets` to list all variable sets applied to a project.
+
 ## Enhancements
 * Update team project access to include additional project roles by @joekarl [#642](https://github.com/hashicorp/go-tfe/pull/642)